% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggquiet.R
\name{create_basic_plot}
\alias{create_basic_plot}
\title{Title create a typical ggplot}
\usage{
create_basic_plot()
}
\value{
a ggplot2 object
}
\description{
Title create a typical ggplot
}
\examples{
create_basic_plot()

}
