% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggquiet.R, R/theme_quiet.R
\name{theme_quiet}
\alias{theme_quiet}
\title{Title all elements to fully transparent}
\usage{
theme_quiet(alpha = 0, color = "black", gridcolor = "white")

theme_quiet(alpha = 0, color = "black", gridcolor = "white")
}
\value{


ggplot object
}
\description{
Title all elements to fully transparent

Title all elements to fully transparent
}
\examples{
create_basic_plot() +
theme_quiet()
g <- create_basic_plot(point_alpha = .2)

g +
theme_quiet(alpha = .1) +
theme_x_axis_loud()
}
