% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_x_axis_quiet.R
\name{theme_y_axis_loud}
\alias{theme_y_axis_loud}
\title{Title}
\usage{
theme_y_axis_loud(color = "black", alpha = 1, gridcolor = "white")
}
\arguments{
\item{color}{}
}
\value{

}
\description{
Title
}
\examples{
library(ggplot2)
ggplot(cars) +
aes(x = speed, y = dist) +
geom_point() +
theme_quiet() +
labs(title = "Hello") +
theme_y_axis_loud()
}
