% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_x_axis_quiet.R
\name{theme_y_axis_quiet}
\alias{theme_y_axis_quiet}
\title{Title}
\usage{
theme_y_axis_quiet(color = "black", alpha = 0, gridcolor = "white")
}
\arguments{
\item{color}{}
}
\value{

}
\description{
Title
}
\examples{
library(ggplot2)
ggplot(cars) +
aes(x = speed, y = dist) +
geom_point() +
theme_y_axis_quiet()
}
