% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggquiet.R
\name{theme_quiet}
\alias{theme_quiet}
\title{Title all elements to fully transparent}
\usage{
theme_quiet()
}
\description{
Title all elements to fully transparent
}
\examples{
create_basic_plot() +
theme_quiet()
}
